+++ src/remote-server/other/index.umd.js
@@ -1784,7 +1784,7 @@ at step (C:\\Users\\me\\work\\webgpt\\promptbook-studio\\node_modules\\node_modu
    * Converts promptbook in JSON format to string format
    *
    * @deprecated TODO: [🥍][🧠] Backup original files in `PipelineJson` same as in Promptbook.studio
-   * @param pipelineJson Promptbook in JSON format (.book.json)
+   * @param pipelineJson Promptbook in JSON format (.bookc)
    * @returns Promptbook in string format (.book.md)
    * @public exported from `@promptbook/core`
    */